Expunge evidence of being a dirty copy-paster
